requirements.txt
aerich
tortoise-orm
sanic